requirements.txt
numpy
matplotlib
pydicom
nibabel
pynrrd
pyyaml
sklearn
scikit-image
scipy
pandas
keras
tensorflow
SimpleITK
nibabel
opencv-python